WIP: changing telemetry client to UDP
The Intel Aero code has not been change to use this changed
client (ie. it won't work until adapted).

Currently the telemetryAPI needs to have the flight_id and
the encryption key passed in. The existing Aero code doesn't
do the startComms() call. Either it will have to do that, or
this telemetry client has to do it.